+++ RELEASE_NOTES.md
@@ -1,7 +1,5 @@
 ### New in 0.39 (not released yet)
 
-* Breaking: `IEventFlowOptions.AddDefaults(...)` now also scans for any
-  aggregate roots
 * New: Autofac is no longer IL merged into the `EventFlow` core NuGet package.
   This is both in preparation for .NET Core and to simplify the build process.
   EventFlow now ships with a custom IoC container by default. The Autofac based
@@ -12,8 +10,11 @@
   with the aggregate ID as argument. The new default also checks if any
   additional services are required for the constructor
 * New: `Command<,,>` now inherits from `ValueObject`
-* Obsolete: `UseResolverAggregateRootFactory()` and `UseResolverAggregateRootFactory()`
+* Obsolete: `UseResolverAggregateRootFactory()` and `UseAutofacAggregateRootFactory()`
   are marked as obsolete as this is now the default
+* Obsolete: All `IEventFlowOptions.AddAggregateRoots(...)` overloads are obsolete,
+  the aggregate factory no longer has any need for the aggregate types to be
+  registered with the container
 
 ### New in 0.38.2454 (released 2016-12-02)